Load language from user's preferences
