+++ app/layout.tsx
@@ -2,14 +2,11 @@ import "./globals.css";
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 
-import { PageContextProvider } from "@contexts/pageContext";
-import { FlightContextProvider } from "@contexts/flightContext";
-import { DialogueContextProvider } from "@contexts/dialogueContext";
 import { RouteChangeListener } from "@components/RouteChangeListener";
 import { Dialogue } from "@components/Dialogue";
 import { IncomingData } from "@components/IncomingData";
 import { Experience } from "@components/Experience";
-import { I18nProvider } from "./i18n";
+import { WithProviders } from "./WithProviders";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -17,19 +14,6 @@ export const metadata: Metadata = {
   title: "Jake Pitman",
   description: "Jake Pitman's personal portfolio website",
 };
-
-type WithProvidersProps = {
-  children: React.ReactNode;
-};
-const WithProviders = ({ children }: WithProvidersProps) => (
-  <PageContextProvider>
-    <FlightContextProvider>
-      <I18nProvider locale="en-us">
-        <DialogueContextProvider>{children}</DialogueContextProvider>
-      </I18nProvider>
-    </FlightContextProvider>
-  </PageContextProvider>
-);
 
 export default function RootLayout({
   children,
